Changes to docker file and prepare file

+++ src/prepare.py
@@ -3,6 +3,7 @@ import pandas as pd
 from io import StringIO
 import sys
 import logging
+import numpy as np
 
 logging.basicConfig(
     format='%(asctime)s %(levelname)s:%(name)s: %(message)s',
@@ -24,10 +25,9 @@ movie_data = pd.read_csv(StringIO(movie_data_path))
 fin_data = pd.read_csv(StringIO(finantial_data_path))
 opening_data = pd.read_csv(StringIO(opening_data_path))
 
-numeric_columns_mask = (movie_data.dtypes == float) | (movie_data.dtypes == int)
+numeric_columns_mask = (movie_data.dtypes == float) | (movie_data.dtypes == np.int64)
 numeric_columns = [column for column in numeric_columns_mask.index if numeric_columns_mask[column]]
 movie_data = movie_data[numeric_columns+['movie_title']]
-
 fin_data = fin_data[['movie_title', 'production_budget', 'worldwide_gross']]
 
 fin_movie_data = pd.merge(fin_data, movie_data, on='movie_title', how='left')
